+++ src/l2cap_signaling.c
@@ -79,22 +79,6 @@ NULL,    // 0x11 non-supported AMP command
 
 static const unsigned int num_l2cap_commands = sizeof(l2cap_signaling_commands_format) / sizeof(const char *);
 
-uint8_t   sig_seq_nr  = 0xff;
-uint16_t  source_cid  = 0x40;
-
-uint8_t l2cap_next_sig_id(void){
-    if (sig_seq_nr == 0xff) {
-        sig_seq_nr = 1;
-    } else {
-        sig_seq_nr++;
-    }
-    return sig_seq_nr;
-}
-
-uint16_t l2cap_next_local_cid(void){
-    return source_cid++;
-}
-
 static uint16_t l2cap_create_signaling_internal(uint8_t * acl_buffer, hci_con_handle_t handle, uint16_t cid, L2CAP_SIGNALING_COMMANDS cmd, uint8_t identifier, va_list argptr){
     
     const char *format = NULL;